Init Redux and for now don't use TypeScript for main entry
